Trying to fix goddamn attacks

+++ LastMinuteJam/Assets/Scripts/Gameplay/Network/HealthIsZeroN.cs
@@ -16,7 +16,7 @@ namespace Platformer.Gameplay
 
         public override void Execute()
         {
-            Schedule<PlayerDeathN>(time);
+            Schedule<PlayerDeathN>((int)time);
         }
     }
 }
