+++ LastMinuteJam/Assets/Scripts/Gameplay/Network/PlayerEnteredDeathZoneN.cs
@@ -17,7 +17,7 @@ namespace Platformer.Gameplay
 
         public override void Execute()
         {
-            SimulationNetick.Schedule<PlayerDeathN>(time);
+            SimulationNetick.Schedule<PlayerDeathN>((int)time);
         }
     }
 }
